tools/feature-checker/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "feature-checker"
version = "1.0.0"
description = "Automated health checks for demo environments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "TestBox TSA Team", email = "tsa@testbox.com"}
]
keywords = ["health-check", "testing", "automation", "playwright", "demo"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-playwright>=0.4.0",
    "black",
    "ruff",
]

[project.scripts]
feature-checker = "feature_checker.cli:main"

[project.urls]
Homepage = "https://github.com/thiagotbx123/tbx-feature-checker"
Documentation = "https://github.com/thiagotbx123/tbx-feature-checker/docs"
Repository = "https://github.com/thiagotbx123/tbx-feature-checker"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
